+++ src/memory/kmalloc.c
@@ -3,7 +3,7 @@
 #include "../util.h"
 #include "../io/log.h"
 #include <stdint.h>
-static uint32_t heap_size;
+static uint64_t heap_size;
 static uint8_t kmalloc_initialized = 0;
 
 //NOTE: This allocator does currently not care about returning allocated pages to the page manager since I don't wanna deal with it
@@ -11,9 +11,9 @@ memory_block_t* head = 0;
 
 void alloc_and_map_new_page(){
     if(heap_size + MEMORY_PAGE_SIZE > KERNEL_MALLOC_END) {panic("Kernel heap has run out of memory"); return;}
-    uint32_t phys_addr = pmm_alloc_page_frame();
+    uint64_t phys_addr = pmm_alloc_page_frame();
     uint32_t n_allocated_pages = CEIL_DIV(heap_size,MEMORY_PAGE_SIZE);
-    uint32_t new_page_addr = KERNEL_MALLOC_START + n_allocated_pages * MEMORY_PAGE_SIZE;
+    uint64_t new_page_addr = KERNEL_MALLOC_START + n_allocated_pages * MEMORY_PAGE_SIZE;
     mem_map_page(new_page_addr,phys_addr,PAGE_FLAG_WRITE);
     heap_size += MEMORY_PAGE_SIZE;
 }
